fix(qsa): 64-bit addressing in the sparse gather kernel
`_compact_kv` formed the pool element offset as `slots * heads * dim` with
int32 `slots` from req_to_token, so it wraps once the pool holds more than
2^31 / (heads * dim) tokens: 4,194,304 for Qwen3.8-Flash-Next (2 x 256).
A BF16 KV pool on one GB300 holds ~3.9M tokens and never reaches that; the FP8
pool holds ~7.6M and does, as soon as the page allocator's high-water mark
crosses 2^16 pages, which in practice happens when the first request
approaches 64K tokens. From then on the gather reads wrong addresses
(garbage/NaN, or an illegal address beyond the pool), poisoning every request
that lands on high slots. This is the FP8-only collapse that survived the
zero-fill fix. Compute the offsets in int64 and add a >2^22-slot regression
test.

+++ python/sglang/srt/layers/attention/qsa/sparse_attn.py
@@ -402,8 +402,15 @@ def _compact_kv(
         mask=valid,
         other=0,
     )
-    src = slots[:, None] * heads * dim + head * dim + dims[None, :]
-    dst = (pack_start + cols)[:, None] * heads * dim + head * dim + dims[None, :]
+    # 64-bit element offsets: slot * heads * dim exceeds int32 once the pool holds
+    # more than 2^31 / (heads * dim) tokens (~4.2M for 2 x 256), which an FP8 pool
+    # on one GPU does reach.
+    src = slots.to(tl.int64)[:, None] * heads * dim + head * dim + dims[None, :]
+    dst = (
+        (pack_start + cols).to(tl.int64)[:, None] * heads * dim
+        + head * dim
+        + dims[None, :]
+    )
     load_mask = valid[:, None] & (dims[None, :] < dim)
     if ZERO_FILL:
         # Strided (page-aligned) packing: the paged decode kernel reads whole pages,
